requirements.txt
certifi==2017.4.17
chardet==3.0.4
idna==2.5
jsonpickle==0.9.4
oauthlib==2.0.2
PyYAML==3.12
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
tweepy==3.5.0
urllib3==1.21.1
